Moves bookmarks related files from js to app folder, split components to a single file per component
Resolves #8351

Auditors: @bsclifton

Test Plan:
- test should be green
- everything should be the same
