+++ app/renderer/components/navigation/navigationBar.js
@@ -15,7 +15,7 @@ const messages = require('../../../../js/constants/messages')
 const settings = require('../../../../js/constants/settings')
 const ipc = require('electron').ipcRenderer
 const {isSourceAboutUrl} = require('../../../../js/lib/appUrlUtil')
-const AddEditBookmarkHanger = require('../addEditBookmarkHanger')
+const AddEditBookmarkHanger = require('../bookmarks/addEditBookmarkHanger')
 const siteUtil = require('../../../../js/state/siteUtil')
 const eventUtil = require('../../../../js/lib/eventUtil')
 const UrlUtil = require('../../../../js/lib/urlutil')
