+++ js/components/main.js
@@ -32,10 +32,10 @@ const ImportBrowserDataPanel = require('../../app/renderer/components/importBrow
 const WidevinePanel = require('../../app/renderer/components/widevinePanel')
 const AutofillAddressPanel = require('./autofillAddressPanel')
 const AutofillCreditCardPanel = require('./autofillCreditCardPanel')
-const AddEditBookmark = require('./addEditBookmark')
+const AddEditBookmark = require('../../app/renderer/components/bookmarks/addEditBookmark')
 const LoginRequired = require('./loginRequired')
 const ReleaseNotes = require('../../app/renderer/components/releaseNotes')
-const BookmarksToolbar = require('../../app/renderer/components/bookmarksToolbar')
+const BookmarksToolbar = require('../../app/renderer/components/bookmarks/bookmarksToolbar')
 const ContextMenu = require('./contextMenu')
 const PopupWindow = require('./popupWindow')
 const NoScriptInfo = require('./noScriptInfo')
